refactor: improve code reuse for event-mappings
