+++ src/event-mappings/alb.js
@@ -1,39 +1,17 @@
-const { mapEventToHttpRequest } = require('./utils')
-const { getEventBody } = require('../utils')
+const { mapEventToHttpRequest, mapResponseToService } = require('./utils')
 
-function mapAlbEventToHttpRequest ({ event, socketPath }) {
-  const httpRequest = mapEventToHttpRequest({ event, socketPath })
-
-  if (event.body) {
-    const body = getEventBody({ event })
-    const isBase64Encoded = event.isBase64Encoded
-    httpRequest.headers['Content-Length'] = Buffer.byteLength(body, isBase64Encoded ? 'base64' : 'utf8')
-  }
-
-  return httpRequest
-}
-
-function mapResponseToAlb ({
+const mapAlbEventToHttpRequest = ({ event, socketPath }) => mapEventToHttpRequest({ event, socketPath })
+const mapResponseToAlb = ({
   statusCode,
   body,
   headers,
   isBase64Encoded
-}) {
-  const multiValueHeaders = {}
-
-  Object.entries(headers).forEach(([headerKey, headerValue]) => {
-    const headerArray = Array.isArray(headerValue) ? headerValue : [headerValue]
-
-    multiValueHeaders[headerKey] = headerArray
-  })
-
-  return {
-    statusCode,
-    body,
-    multiValueHeaders,
-    isBase64Encoded
-  }
-}
+}) => mapResponseToService({
+  statusCode,
+  body,
+  headers,
+  isBase64Encoded
+})
 
 module.exports = {
   mapAlbEventToHttpRequest,
