+++ src/event-mappings/lambda-edge.js
@@ -1,24 +1,17 @@
-const { mapEventToHttpRequest } = require('./utils')
+const { mapEventToHttpRequest, mapResponseToService } = require('./utils')
 
-function mapLambdaEdgeEventToHttpRequest ({ event, socketPath }) {
-  const httpRequest = mapEventToHttpRequest({ event, socketPath })
-
-  return httpRequest
-}
-
-function mapResponseToLambdaEdge ({
+const mapLambdaEdgeEventToHttpRequest = ({ event, socketPath }) => mapEventToHttpRequest({ event, socketPath })
+const mapResponseToLambdaEdge = ({
   statusCode,
   body,
   headers,
   isBase64Encoded
-}) {
-  return {
-    statusCode,
-    body,
-    multiValueHeaders: headers,
-    isBase64Encoded
-  }
-}
+}) => mapResponseToService({
+  statusCode,
+  body,
+  headers,
+  isBase64Encoded
+})
 
 module.exports = {
   mapLambdaEdgeEventToHttpRequest,
